src/requirements.txt
django
djangorestframework
pyyaml
requests